phage_pipeline/requirements_py-env.txt
biopython==1.76
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coloredlogs==10.0
colormath==3.0.0
cycler==0.10.0
decorator==4.4.1
future==0.18.2
humanfriendly==4.18
idna==2.8
Jinja2==2.10.3
kiwisolver==1.1.0
lzstring==1.0.4
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
multiqc==1.8
networkx==2.2
numpy==1.18.1
pandas==0.24.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyVCF==0.6.8
PyYAML==5.3
requests==2.22.0
simplejson==3.17.0
six==1.13.0
spectra==0.0.11
urllib3==1.25.7
